Feed FV1 clock signal from nRF SoC
I think the oscillator in the FV1 is busted. Tried two different crystals,
a whole lot of loading cap/res values and sometimes it will start, and
start for a few days reliably and then miserably stall the next week.

Output a clean clock signal directly from the SoC, noise be damned.
A TODO is add a LPF as close as possible to the SoC.