load sample programs
Load sample programs so the devices does something without having to have a
PC to boot it up.

Programs gotten from the default project from SpinASM. I think they're
supposed to be the ones in the ROM, described in the datasheet.

I can't just point the I2C TX buffer to the ROM arrays as the nRF52 DMA
engine cannot from FLASH, the data has to be in RAM.